requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
coloredlogs==15.0.1
flatbuffers==23.5.26
fusion-solar-py==0.0.15
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
numpy==1.26.1
onnxruntime==1.16.1
opencv-python==4.8.1.78
packaging==23.2
protobuf==4.24.4
pymodbus==2.5.3
pyModbusTCP==0.2.0
pyreadline3==3.4.1
pyserial==3.5
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sun2000_modbus==2.1.0
sympy==1.12
urllib3==2.0.7
